Try to fix watchOS build.
* platform/graphics/Region.h:
(WebCore::Region::Span::decode):



Canonical link: https://commits.webkit.org/210084@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@242998 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/Region.h
@@ -227,7 +227,7 @@ Optional<Region::Span> Region::Span::decode(Decoder& decoder)
     if (!segmentIndex)
         return { };
 
-    return { { *y, *segmentIndex } };
+    return { { *y, static_cast<size_t>(*segmentIndex) } };
 }
 
 template<class Encoder>
